+++ src/orderBy/__tests__/index.test.ts
@@ -12,22 +12,28 @@ describe('orderBy()', () => {
     jest.resetAllMocks();
   });
 
-  it('should call baseOrderBy() with provided collection argument and an empty array value for identifiers and orders', () => {
+  it('should call baseOrderBy() with provided collection argument and an empty array value for identifiers and orders and undefined locale', () => {
     const collection = ['Fred', 'barney', 'frank', 'Bob'];
     const identifiers = undefined;
     const orders = undefined;
 
     orderBy(collection, identifiers, orders);
     expect(baseOrderBy).toHaveBeenCalledTimes(1);
-    expect(baseOrderBy).toHaveBeenCalledWith(collection, [], []);
+    expect(baseOrderBy).toHaveBeenCalledWith(collection, [], [], undefined);
   });
-  it('should call baseOrderBy() with provided collection, identifiers and orders arguments', () => {
+  it('should call baseOrderBy() with provided collection, identifiers, orders and locale arguments', () => {
     const collection = ['Fred', 'barney', 'frank', 'Bob'];
     const identifiers = [(v: string) => v.toLowerCase()];
     const orders: Array<Order> = ['desc'];
+    const locale = 'de';
 
-    orderBy(collection, identifiers, orders);
+    orderBy(collection, identifiers, orders, locale);
     expect(baseOrderBy).toHaveBeenCalledTimes(1);
-    expect(baseOrderBy).toHaveBeenCalledWith(collection, identifiers, orders);
+    expect(baseOrderBy).toHaveBeenCalledWith(
+      collection,
+      identifiers,
+      orders,
+      locale,
+    );
   });
 });