+++ src/utils/baseCompare.ts
@@ -1,6 +1,6 @@
-import type { BaseCompareOptions } from '../types';
 import { compareValues } from './compareValues';
 import { getMappedValueRecord } from './getMappedValueRecord';
+import type { BaseCompareOptions } from '../types';
 
 export const baseCompare =
   (options: BaseCompareOptions) =>
@@ -8,7 +8,7 @@ export const baseCompare =
     const a = getMappedValueRecord(valueA);
     const b = getMappedValueRecord(valueB);
 
-    const result = compareValues(a, b);
+    const result = compareValues(a, b, options.locale);
 
     return result * (options.order === 'desc' ? -1 : 1);
   };