+++ src/utils/baseOrderBy.ts
@@ -1,19 +1,21 @@
-import type {
-  Identifier,
-  Order,
-  MappedCollection,
-  IdentifierFn,
-} from '../types';
 import { compareMultiple } from './compareMultiple';
 import { createIdentifierFn } from './createIdentifierFn';
 import { getElementByIndex } from './getElementByIndex';
 import { getMappedValueRecord } from './getMappedValueRecord';
 import { getValueByIdentifier } from './getValueByIdentifier';
+import type {
+  Identifier,
+  Order,
+  MappedCollection,
+  IdentifierFn,
+  Locale,
+} from '../types';
 
 export const baseOrderBy = <T>(
   collection: ReadonlyArray<T>,
   identifiers: ReadonlyArray<Identifier<T>>,
-  orders: ReadonlyArray<Order>
+  orders: ReadonlyArray<Order>,
+  locale?: Locale,
 ): Array<T> => {
   const identifierFns: Array<IdentifierFn<T>> = identifiers.length
     ? identifiers.map(createIdentifierFn)
@@ -29,15 +31,15 @@ export const baseOrderBy = <T>(
         index,
         values,
       };
-    }
+    },
   );
 
   // iterate over values and compare values until a != b or last value reached
   mappedCollection.sort((recordA, recordB) =>
-    compareMultiple(recordA, recordB, orders)
+    compareMultiple(recordA, recordB, orders, locale),
   );
 
   return mappedCollection.map((element) =>
-    getElementByIndex(collection, element.index)
+    getElementByIndex(collection, element.index),
   );
 };