+++ src/utils/compareValues.ts
@@ -1,11 +1,12 @@
-import type { MappedValueRecord } from '../types';
 import { compareChunks } from './compareChunks';
 import { compareNumbers } from './compareNumbers';
 import { compareOtherTypes } from './compareOtherTypes';
+import type { MappedValueRecord, Locale } from '../types';
 
 export const compareValues = (
   valueA: MappedValueRecord,
-  valueB: MappedValueRecord
+  valueB: MappedValueRecord,
+  locale?: Locale,
 ): number => {
   if (valueA.value === valueB.value) {
     return 0;
@@ -14,7 +15,7 @@ export const compareValues = (
     return compareNumbers(valueA.parsedNumber, valueB.parsedNumber);
   }
   if (valueA.chunks && valueB.chunks) {
-    return compareChunks(valueA.chunks, valueB.chunks);
+    return compareChunks(valueA.chunks, valueB.chunks, locale);
   }
   return compareOtherTypes(valueA, valueB);
 };