+++ src/utils/getOptions.ts
@@ -4,20 +4,22 @@ const isValidOrder = (value: unknown): boolean =>
   typeof value === 'string' && (value === 'asc' || value === 'desc');
 
 export const getOptions = (
-  customOptions?: CompareOptions
+  customOptions?: CompareOptions,
 ): BaseCompareOptions => {
   let order: OrderEnum = 'asc';
+  let locale; // = 'en';
   if (typeof customOptions === 'string' && isValidOrder(customOptions)) {
     order = customOptions;
-  } else if (
-    customOptions &&
-    typeof customOptions === 'object' &&
-    customOptions.order &&
-    isValidOrder(customOptions.order)
-  ) {
-    ({ order } = customOptions);
+  } else if (customOptions && typeof customOptions === 'object') {
+    if (customOptions.order && isValidOrder(customOptions.order)) {
+      order = customOptions.order;
+    }
+    if (customOptions.locale && customOptions.locale.length > 0) {
+      locale = customOptions.locale;
+    }
   }
   return {
     order,
+    locale,
   };
 };