+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # natural-orderby
 
+## 5.0.0
+
+### Major Changes
+
+- feat: support numeric separators
+- feat: add ability to define locale for sorting unicode
+
+## 4.0.0
+
+### Major Changes
+
+- fix: update `Identifier<T>` to be `keyof T` instead of any `string`
+
 ## 3.0.2
 
 ### Patch Changes